Change spinner's options on show call
Use case is for people that wish to use one single spinner throughout the
app, passing different options each call, so they can have one single
instance that behaves as multiple instances.

Given recent improvements on spinner instantiation, that was as easy as
changing some methods signatures and assigning to the spinner's object
those options, when they were specified.

Now we can show a spinner with it's name(or blank, in case of unnamed
spinner), OR it's name and options, that range as long as the Spinner
interface goes i.e. bdColor, size, color, type.

I decided not to let someone change the spinner's name by options, for it's
a constant, would be kind of a redundant implementation and... doesn't make
much sense for the proposed use case.

Note: If you want to change the options for an unnamed spinner it will be
necessary to call show(undefined, options) explicitly, otherwise just use
the name of the spinner in place of undefined.

Closes #8 "i use one spinner for all pages in my project, how to change
options" - Enhancement tag.